pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "retrocode"
version = "0.1.0"
description = "Turn Claude Code session traces into an auto-updating project playbook"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "anthropic>=0.84.0",
    "openai>=2.18.0",
]

[project.scripts]
retro = "src.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
